graphs display working and filtering

+++ memcache_set_vars.py
@@ -8,6 +8,9 @@ from datetime import date
 from datetime import time
 import datetime
 import subprocess
+import syslog
+import os
+
 #conn=Connection('10.245.145.84')
 conn=Connection('10.115.126.7')
 db=conn.wc
@@ -71,7 +74,7 @@ def GenHourlyGraph(id):
 		pass
 	OFILE.close()
 	subprocess.call(["gnuplot","gnuplot.plot"])
-	OUTFILENAME='/tmp/django/wikicount/static/images/'+str(id)+'.png'
+	OUTFILENAME='/tmp/django/wikicount/static/images/hourly/'+str(id)+'.png'
 	SFILE='/tmp/django/wikicount/introduction.png'
 	subprocess.Popen("mv "+str(SFILE)+" "+str(OUTFILENAME),shell=True)
 	return
@@ -100,20 +103,47 @@ def GenInfoPage(id):
 	info_lt5k_list=[]
 	info_lt500_list=[]
 	info_lt50_list=[]
-        for result in DFINDQ:
+        
+	OFILE250K=open("/tmp/t250k.log","w")	
+	for result in DFINDQ:
 	        rec={'d':str(result['d']),'m':str(result['m']),'y':str(result['y']),'place':str(result['place'])}
+		OFILE250K.write(str(rec['y'])+'/'+str(rec['m'])+'/'+str(rec['d'])+' '+str(rec['place'])+'\n')
         	send_list.append(rec)
         for result in FINDQ:
 	        rec={'d':str(result['d']),'m':str(result['m']),'y':str(result['y']),'place':str(result['place'])}
+		OFILE250K.write(str(rec['y'])+'/'+str(rec['m'])+'/'+str(rec['d'])+' '+str(rec['place'])+'\n')
         	send_list.append(rec)
-	
+	OFILE250K.close()
+	OUTFILENAME='/tmp/django/wikicount/static/images/t250k/'+str(id)+'.png'
+	if os.path.lexists(OUTFILENAME) and random.randint(0,20)==10:
+		subprocess.call(["gnuplot","gnuplot.250k"])
+		SFILE='/tmp/t250k.png'
+		subprocess.Popen("mv "+str(SFILE)+" "+str(OUTFILENAME),shell=True)
+	elif not os.path.lexists(OUTFILENAME):
+		subprocess.call(["gnuplot","gnuplot.250k"])
+		SFILE='/tmp/t250k.png'
+		subprocess.Popen("mv "+str(SFILE)+" "+str(OUTFILENAME),shell=True)
+			
+	OFILE50K=open("/tmp/t50k.log","w")	
 	LT50KQ=db[thCN].find(Q50K)
         for result in D50KFINDQ:
 	        rec={'d':str(result['d']),'m':str(result['m']),'y':str(result['y']),'place':str(result['place'])}
+		OFILE50K.write(str(rec['y'])+'/'+str(rec['m'])+'/'+str(rec['d'])+' '+str(rec['place'])+'\n')
         	info_lt50k_list.append(rec)
         for result in LT50KQ:
 	        rec={'d':str(result['d']),'m':str(result['m']),'y':str(result['y']),'place':str(result['place'])}
+		OFILE50K.write(str(rec['y'])+'/'+str(rec['m'])+'/'+str(rec['d'])+' '+str(rec['place'])+'\n')
         	info_lt50k_list.append(rec)
+	OFILE50K.close()
+	OUTFILENAME='/tmp/django/wikicount/static/images/t50k/'+str(id)+'.png'
+	if os.path.lexists(OUTFILENAME) and random.randint(0,20)==10:
+		subprocess.call(["gnuplot","gnuplot.50k"])
+		SFILE='/tmp/t50k.png'
+		subprocess.Popen("mv "+str(SFILE)+" "+str(OUTFILENAME),shell=True)
+	elif not os.path.lexists(OUTFILENAME):
+		subprocess.call(["gnuplot","gnuplot.50k"])
+		SFILE='/tmp/t50k.png'
+		subprocess.Popen("mv "+str(SFILE)+" "+str(OUTFILENAME),shell=True)
 
 	OFILE5K=open("/tmp/t5k.log","w")	
 	LT5KQ=db[thCN].find(Q5K)
@@ -126,10 +156,15 @@ def GenInfoPage(id):
 		OFILE5K.write(str(rec['y'])+'/'+str(rec['m'])+'/'+str(rec['d'])+' '+str(rec['place'])+'\n')
         	info_lt5k_list.append(rec)
 	OFILE5K.close()
-	subprocess.call(["gnuplot","gnuplot.5k"])
-	OUTFILENAME='/tmp/django/wikicount/static/images/t5k_'+str(id)+'.png'
-	SFILE='/tmp/t5k.png'
-	subprocess.Popen("mv "+str(SFILE)+" "+str(OUTFILENAME),shell=True)
+	OUTFILENAME='/tmp/django/wikicount/static/images/t5k/'+str(id)+'.png'
+	if os.path.lexists(OUTFILENAME) and random.randint(0,20)==10:
+		subprocess.call(["gnuplot","gnuplot.5k"])
+		SFILE='/tmp/t5k.png'
+		subprocess.Popen("mv "+str(SFILE)+" "+str(OUTFILENAME),shell=True)
+	elif not os.path.lexists(OUTFILENAME):
+		subprocess.call(["gnuplot","gnuplot.5k"])
+		SFILE='/tmp/t5k.png'
+		subprocess.Popen("mv "+str(SFILE)+" "+str(OUTFILENAME),shell=True)
 
 	LT500=db[thCN].find(Q500)
 	OFILE500=open("/tmp/top500.log","w")
@@ -142,18 +177,38 @@ def GenInfoPage(id):
 		OFILE500.write(str(rec['y'])+'/'+str(rec['m'])+'/'+str(rec['d'])+' '+str(rec['place'])+'\n')
         	info_lt500_list.append(rec)
 	OFILE500.close()	
-	subprocess.call(["gnuplot","gnuplot.500"])
-	OUTFILENAME='/tmp/django/wikicount/static/images/t500_'+str(id)+'.png'
-	SFILE='/tmp/top500.png'
-	subprocess.Popen("mv "+str(SFILE)+" "+str(OUTFILENAME),shell=True)
+	OUTFILENAME='/tmp/django/wikicount/static/images/t500/'+str(id)+'.png'
+	if os.path.lexists(OUTFILENAME) and random.randint(0,20)==10:
+		subprocess.call(["gnuplot","gnuplot.500"])
+		SFILE='/tmp/top500.png'
+		subprocess.Popen("mv "+str(SFILE)+" "+str(OUTFILENAME),shell=True)
+	elif not os.path.lexists(OUTFILENAME):
+		subprocess.call(["gnuplot","gnuplot.500"])
+		SFILE='/tmp/top500.png'
+		subprocess.Popen("mv "+str(SFILE)+" "+str(OUTFILENAME),shell=True)
 
 	LT50=db[thCN].find(Q50)
+	OFILE50=open("/tmp/top50.log","w")
         for result in D50FINDQ:
 	        rec={'d':str(result['d']),'m':str(result['m']),'y':str(result['y']),'place':str(result['place'])}
+		OFILE50.write(str(rec['y'])+'/'+str(rec['m'])+'/'+str(rec['d'])+' '+str(rec['place'])+'\n')
         	info_lt50_list.append(rec)
         for result in LT50:
 	        rec={'d':str(result['d']),'m':str(result['m']),'y':str(result['y']),'place':str(result['place'])}
+		OFILE50.write(str(rec['y'])+'/'+str(rec['m'])+'/'+str(rec['d'])+' '+str(rec['place'])+'\n')
         	info_lt50_list.append(rec)
+	OFILE50.close()	
+	OUTFILENAME='/tmp/django/wikicount/static/images/t50/'+str(id)+'.png'
+	if os.path.lexists(OUTFILENAME) and random.randint(0,20)==10:
+		subprocess.call(["gnuplot","gnuplot.50"])
+		SFILE='/tmp/top50.png'
+		subprocess.Popen("mv "+str(SFILE)+" "+str(OUTFILENAME),shell=True)
+	elif not os.path.lexists(OUTFILENAME):
+		subprocess.call(["gnuplot","gnuplot.50"])
+		SFILE='/tmp/top50.png'
+		subprocess.Popen("mv "+str(SFILE)+" "+str(OUTFILENAME),shell=True)
+
+
         mc.set(INFOVIEW_KEY,send_list,60*60*12)
 	mc.set(INFOVIEWLT_KEY,info_lt50k_list,60*60*12)
 	mc.set(INFOVIEWLT5K_KEY,info_lt5k_list,60*60*12)
@@ -171,9 +226,10 @@ mc.set('mcdpDaysList',send_list,60*60*24)
 
 
 
-QUERY={'d':int(DAY),'m':int(MONTH),'y':int(YEAR)}
+QUERY={'d':DAY,'m':MONTH,'y':YEAR}
 send_list=[]
 RESULTSET=db[thCN].find(QUERY).sort('place',1).limit(100)
+syslog.syslog('memcache-daily: '+str(QUERY)+' count: '+str(RESULTSET.count()))
 for row in RESULTSET:
 	GenInfoPage(row['id'])
 	title,utitle=FormatName(row['title'])
@@ -183,6 +239,7 @@ mc.set('DAYKEY',send_list,7200)
 notedate=''
 notes=''
 latest_hits_list = db[thCN].find(QUERY).sort('place',1).limit(100)
+syslog.syslog('memcache-latest: '+str(QUERY)+' count: '+str(latest_hits_list.count()))
 for p in latest_hits_list:
 	GenInfoPage(p['id'])
         title,utitle=FormatName(p['title'])
@@ -191,9 +248,10 @@ for p in latest_hits_list:
 mc.set('send_list',send_list,3600)
 
 
-print 'now trending list query....'
 send_list=[]    
-TRENDING_LIST_QUERY=db.prodtrend.find({'d':int(DAY),'m':int(MONTH),'y':int(YEAR)}).sort('Hits',-1).limit(150)
+TRENDING_LIST_QUERY=db.prodtrend.find({u'd':DAY,u'm':MONTH,u'y':YEAR}).sort('Hits',-1).limit(150)
+syslog.syslog('memcache-trending: '+str(QUERY)+' count: '+str(TRENDING_LIST_QUERY.count()))
+print TRENDING_LIST_QUERY.count()
 for p in TRENDING_LIST_QUERY:
 	rec={'title':p['title'],'place':p['place'],'Hits':p['Hits']%1000,'linktitle':p['linktitle'],'id':p['id']}
         send_list.append(rec)
@@ -205,7 +263,7 @@ print 'random query...'
 send_list=[]
 for a in range(1,50):
 	place=random.randint(1,250000)
-        FQUERY={'d':int(DAY),'m':int(MONTH),'y':int(YEAR),'place':place}
+        FQUERY={u'd':DAY,u'm':MONTH,u'y':YEAR,'place':place}
         RANDOM_LIST_QUERY=db[thCN].find(FQUERY)
         for item in RANDOM_LIST_QUERY:
         	 id=item['id']
@@ -221,7 +279,8 @@ print 'debuts query...'
 send_list=[]
 title=''
 utitle=''
-QUERY=db[dbCN].find({'d':int(DAY),'m':int(MONTH),'y':int(YEAR)}).sort('place',1).limit(300)
+QUERY=db[dbCN].find({u'd':DAY,u'm':MONTH,u'y':YEAR}).sort('place',1).limit(300)
+syslog.syslog('memcache-debuts: count: '+str(QUERY.count()))
 for item in QUERY:
 	COUNT=0
         TITLE=''
@@ -238,13 +297,15 @@ send_list=[]
 RESULTSET=db.command({'distinct':thCN,'key':'d','query':{'m':int(MONTH),'y':int(YEAR)}})
 for d in RESULTSET['values']:
 	rec={'d':d,'m':MONTH,'y':YEAR,'stry':str(YEAR),'strm':str(MONTH),'strd':str(d)}
-	QUERY={'d':int(DAY),'m':int(MONTH),'y':int(YEAR)}
+	QUERY={'d':int(d),'m':int(MONTH),'y':int(YEAR)}
 	DAYKEY='toplist'+str(YEAR)+str(MONTH)+str(DAY)
 	page_list=[]
         PAGERESULTSET=db[thCN].find(QUERY).sort('place',1).limit(100)
+	syslog.syslog('memcache-monthly: '+str(d)+' '+str(QUERY)+' count: '+str(PAGERESULTSET.count()))
         for row in PAGERESULTSET:
 		title, utitle=FormatName(row['title'])
                 prec={'place':row['place'],'Hits':row['Hits'],'title':title ,'id':str(row['id']),'linktitle':utitle}
+		GenInfoPage(row['id'])
                 page_list.append(prec)
                 mc.set('DAYKEY',page_list,60*60*24*14)
         send_list.append(rec)
@@ -270,6 +331,7 @@ mc.set('mcdpDaysList'+str(MONTH)+str(YEAR),send_list,60*60*24)
 print 'lastly, 3hr rolling average'
 send_list=[]
 THREEHOUR_LIST_QUERY=db.threehour.find().sort('place',1)
+syslog.syslog('memcache-threehour:  count: '+str(THREEHOUR_LIST_QUERY.count()))
 for p in THREEHOUR_LIST_QUERY:
 	rec={'title':p['title'],'place':p['place'],'Avg':p['rollavg'],'linktitle':p['title'],'id':p['id']}
 	GenInfoPage(p['id'])
@@ -283,6 +345,7 @@ HOURQUERY=db.hitshourlydaily.find({str(SEARCH_HOUR):{'$gt':1}}).sort(str(SEARCH_
 send_list=[]
 place=1
 HOURKEY="SEARCHHOUR_"+str(SEARCH_HOUR)
+syslog.syslog('memcache-hourly: '+' count: '+str(HOURQUERY.count()))
 for row in HOURQUERY:
     title,utitle=MapQuery_FindName(row['_id'])
     GenInfoPage(row['_id'])
